+++ macros/src/lib.rs
@@ -17,12 +17,13 @@ use crate::{deps::Dependencies, utils::format_generics};
 mod utils;
 mod attr;
 mod deps;
+mod optional;
 mod types;
 
 struct DerivedTS {
     crate_rename: Path,
     ts_name: Expr,
-    docs: String,
+    docs: Vec<Expr>,
     inline: TokenStream,
     inline_flattened: Option<TokenStream>,
     dependencies: Dependencies,
@@ -63,12 +64,15 @@ impl DerivedTS {
             }
         };
 
-        let docs = match &*self.docs {
-            "" => None,
-            docs => Some(quote!(const DOCS: Option<&'static str> = Some(#docs);)),
-        };
-
         let crate_rename = self.crate_rename.clone();
+        let docs = match &*self.docs {
+            [] => None,
+            docs => Some(quote! {
+                fn docs() -> Option<String> {
+                    Some(#crate_rename::format_docs(&[#(#docs),*]))
+                }
+            }),
+        };
 
         let ident = self.ts_name.clone();
         let impl_start = generate_impl_block_header(