+++ macros/src/types/named.rs
@@ -1,10 +1,11 @@
 use proc_macro2::TokenStream;
-use quote::{quote, quote_spanned};
-use syn::{parse_quote, spanned::Spanned, Expr, Field, FieldsNamed, Path, Result};
+use quote::quote;
+use syn::{spanned::Spanned, Expr, Field, FieldsNamed, Path, Result};
 
 use crate::{
-    attr::{Attr, ContainerAttr, FieldAttr, Inflection, Optional, StructAttr},
+    attr::{Attr, ContainerAttr, FieldAttr, Inflection, StructAttr},
     deps::Dependencies,
+    optional::Optional,
     utils::{raw_name_to_ts_field, to_ts_ident},
     DerivedTS,
 };
@@ -101,58 +102,16 @@ fn format_field(
         return Ok(());
     }
 
-    if let Some(ref type_override) = field_attr.type_override {
-        let field_name = to_ts_ident(field.ident.as_ref().unwrap());
-        let name = match (field_attr.rename.as_ref(), rename_all) {
-            (Some(rn), _) => rn.to_owned(),
-            (None, Some(rn)) => rn.apply(&field_name),
-            (None, None) => field_name,
-        };
-        let valid_name = raw_name_to_ts_field(name);
-
-        // Start every doc string with a newline, because when other characters are in front, it is not "understood" by VSCode
-        let docs = match field_attr.docs.is_empty() {
-            true => "".to_string(),
-            false => format!("\n{}", &field_attr.docs),
-        };
-
-        formatted_fields.push(quote! {
-            format!("{}{}: {},", #docs, #valid_name, #type_override)
-        });
-
-        return Ok(());
-    }
-
     let ty = field_attr.type_as(&field.ty);
 
-    let (optional_annotation, nullable) = match (struct_optional, field_attr.optional) {
-        // `#[ts(optional)]` on field takes precedence, and is enforced **AT COMPILE TIME**
-        (_, Optional::Optional { nullable }) => (
-            // expression that evaluates to the string "?", but fails to compile if `ty` is not an `Option`.
-            quote_spanned! { field.span() => {
-                fn check_that_field_is_option<T: #crate_rename::IsOption>(_: std::marker::PhantomData<T>) {}
-                let x: std::marker::PhantomData<#ty> = std::marker::PhantomData;
-                check_that_field_is_option(x);
-                "?"
-            }},
-            nullable,
-        ),
-        // `#[ts(optional)]` on the struct acts as `#[ts(optional)]` on a field, but does not error on non-`Option`
-        // fields. Instead, it is a no-op.
-        (Optional::Optional { nullable }, _) => (
-            quote! {
-                if <#ty as #crate_rename::TS>::IS_OPTION { "?" } else { "" }
-            },
-            nullable,
-        ),
-        _ => (quote!(""), true),
-    };
-
-    let ty = if nullable {
-        ty
-    } else {
-        parse_quote! {<#ty as #crate_rename::TS>::OptionInnerType}
-    };
+    let (is_optional, ty) = crate::optional::apply(
+        crate_rename,
+        struct_optional,
+        &ty,
+        &field_attr,
+        field.span(),
+    );
+    let optional_annotation = quote!(if #is_optional { "?" } else { "" });
 
     if field_attr.flatten {
         flattened_fields.push(quote!(<#ty as #crate_rename::TS>::inline_flattened()));
@@ -160,13 +119,18 @@ fn format_field(
         return Ok(());
     }
 
-    let formatted_ty = if field_attr.inline {
-        dependencies.append_from(&ty);
-        quote!(<#ty as #crate_rename::TS>::inline())
-    } else {
-        dependencies.push(&ty);
-        quote!(<#ty as #crate_rename::TS>::name())
-    };
+    let formatted_ty = field_attr
+        .type_override
+        .map(|t| quote!(#t))
+        .unwrap_or_else(|| {
+            if field_attr.inline {
+                dependencies.append_from(&ty);
+                quote!(<#ty as #crate_rename::TS>::inline())
+            } else {
+                dependencies.push(&ty);
+                quote!(<#ty as #crate_rename::TS>::name())
+            }
+        });
 
     let field_name = to_ts_ident(field.ident.as_ref().unwrap());
     let name = match (field_attr.rename, rename_all) {
@@ -177,9 +141,9 @@ fn format_field(
     let valid_name = raw_name_to_ts_field(name);
 
     // Start every doc string with a newline, because when other characters are in front, it is not "understood" by VSCode
-    let docs = match field_attr.docs.is_empty() {
-        true => "".to_string(),
-        false => format!("\n{}", &field_attr.docs),
+    let docs = match &*field_attr.docs {
+        &[] => quote!(""),
+        docs => quote!(format!("\n{}", #crate_rename::format_docs(&[#(#docs),*]))),
     };
 
     formatted_fields.push(quote! {